Adding incoming packet parsing to the demo program

+++ guitest.py
@@ -12,7 +12,7 @@ if __name__ == '__main__':
     dt4 = manager.DataType('test4', float)
     dt5 = manager.DataType('test5', str, False)
     dt6 = manager.DataType('test6', float)
-    dt7 = manager.DataType('test7', float)
+    dt7 = manager.DataType('test7', int)
     dt8 = manager.DataType('test8', float)
     dispatcher = manager.Dispatcher(dt1, dt2, dt3, dt4, dt5, dt6, dt7, dt8)
     manager = manager.DataManager(dispatcher)
